+++ R/tar_target.R
@@ -151,7 +151,10 @@
 #'   * `"list"`, branching happens with `[[]]` and aggregation happens with
 #'     `list()`.
 #'   * `"group"`: `dplyr::group_by()`-like functionality to branch over
-#'     subsets of a data frame. The target's return value must be a data
+#'     subsets of a non-dynamic data frame.
+#'     For `iteration = "group"`, the target must not by dynamic
+#'     (the `pattern` argument of [tar_target()] must be left `NULL`).
+#'     The target's return value must be a data
 #'     frame with a special `tar_group` column of consecutive integers
 #'     from 1 through the number of groups. Each integer designates a group,
 #'     and a branch is created for each collection of rows in a group.